Moved all enums to enums package Reorganized all imports
